+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * added in epl
 * added in pulling betting records for a season for teams/league
 * fixed pulling team stats
-* 
+* removed client functions that are now found within other classes. i.e. betting calls are available in teams
 
 ## 0.3.1
 * added function to client/team to pull a seasons roster