+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## 0.3.2
 * added in epl
-* added in pulling betting records for a season for teams/league
+* added in pulling betting records for a season for teams/ teams in a league
 * fixed pulling team stats
 * removed client functions that are now found within other classes. i.e. betting calls are available in teams
+* 
 
 ## 0.3.1
 * added function to client/team to pull a seasons roster